pyproject.toml
[project]
name = "wundernn-challenge"
version = "0.1.0"
description = "Market next state prediction"
authors = [
    {name = "alexnachyn@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.15"
dependencies = [
    "pandas (>=2.3.3,<3.0.0)",
    "scikit-learn (>=1.7.2,<2.0.0)",
    "pyarrow (>=22.0.0,<23.0.0)",
    "numpy (>=2.3.4,<3.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "matplotlib (>=3.10.7,<4.0.0)",
    "statsmodels (>=0.14.5,<0.15.0)",
    "torch (>=2.9.0,<3.0.0)",
    "torchvision (>=0.24.0,<0.25.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "lightning (>=2.5.5,<3.0.0)",
    "tensorboardx (>=2.6.4,<3.0.0)",
    "tensorboard (>=2.20.0,<3.0.0)",
    "scipy (>=1.16.3,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
